MCR-1618 extract internal classes
