+++ mycore-oai/src/main/java/org/mycore/oai/set/MCROAIClassificationToSetHandler.java
@@ -2,11 +2,7 @@ package org.mycore.oai.set;
 
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -16,11 +12,9 @@ import org.apache.solr.client.solrj.response.QueryResponse;
 import org.apache.solr.common.SolrDocument;
 import org.apache.solr.common.params.CommonParams;
 import org.apache.solr.common.params.ModifiableSolrParams;
-import org.mycore.common.MCRException;
 import org.mycore.common.config.MCRConfiguration;
 import org.mycore.oai.MCROAIUtils;
 import org.mycore.oai.classmapping.MCRClassificationAndSetMapper;
-import org.mycore.oai.pmh.Set;
 import org.mycore.solr.MCRSolrClientFactory;
 import org.mycore.solr.MCRSolrUtils;
 
@@ -91,52 +85,6 @@ public class MCROAIClassificationToSetHandler extends MCROAISolrSetHandler {
         MCROAIClassificationSetResolver resolver = new MCROAIClassificationSetResolver();
         resolver.init(getConfigPrefix(), getHandlerPrefix(), getSetMap(), result, MCROAISolrSetHandler::getIdentifier);
         return resolver;
-    }
-
-    protected static class MCROAIClassificationSetResolver extends MCROAISetResolver<String, SolrDocument> {
-
-        Map<String, SolrDocument> setMap;
-
-        private String classId;
-
-        private String classField;
-
-        private String classPrefix;
-
-        @Override
-        public void init(String configPrefix, String setId, Map<String, MCRSet> setMap, Collection<SolrDocument> result,
-            Function<SolrDocument, String> identifier) {
-            super.init(configPrefix, setId, setMap, result, identifier);
-            this.setMap = result.stream().collect(Collectors.toMap(getIdentifier(), d -> d));
-            classId = MCRClassificationAndSetMapper.mapSetToClassification(configPrefix, setId);
-            classField = MCRConfiguration.instance().getString(getConfigPrefix() + "SetSolrField", "category.top");
-            classPrefix = classId + ":";
-        }
-
-        @Override
-        public Collection<Set> getSets(String key) {
-            SolrDocument solrDocument = setMap.get(key);
-            if (solrDocument == null) {
-                throw new MCRException("Unknown key: " + key);
-            }
-            return Optional.ofNullable(solrDocument.getFieldValues(classField))
-                .orElseGet(() -> Collections.emptySet())
-                .stream()
-                .map(String.class::cast)
-                .filter(s -> s.startsWith(classPrefix))
-                .map(s -> s.substring(classPrefix.length()))
-                .map(s -> getSetId() + ":" + s)
-                .map(s -> {
-                    Set set = getSetMap().get(s);
-                    if (set == null) {
-                        throw new MCRException("Set " + s + " is undefined: " + getSetMap());
-                    }
-                    return set;
-                })
-                .collect(Collectors.toSet());
-
-        }
-
     }
 
 }
